Move env var logic to a better home
